섹션 22-259 완료
적 처치 시 XP 증가 구현
HUD에 XP 표시하는 기능 구현
Interface 사용 구문 일관화

+++ Source/Aura/UI/WidgetController/OverlayWidgetController.cpp
@@ -108,7 +108,7 @@ void UOverlayWidgetController::OnXPChanged(const int32 InXP) const
 		const int32 DeltaLevelUpRequirement = LevelUpRequirement - PreviousLevelUpRequirement;
 		const int32 XPForThisLevel = InXP - PreviousLevelUpRequirement;
 
-		const float XPBarPercent = static_cast<float>(XPForThisLevel / DeltaLevelUpRequirement);
+		const float XPBarPercent = static_cast<float>(XPForThisLevel) / static_cast<float>(DeltaLevelUpRequirement);
 		OnXPBarPercentChangedDelegate.Broadcast(XPBarPercent);
 	}
 }
